bloaty: fix signed integer overflow due to use of (s)size_t types
Fix signed integer overflow in the 32-bit version of bloaty when
analyzing libraries with a combined size of >2GB, due to the use of
ssize_t (and size_t) when int64_t should have been used.

Before this change:

$ bloaty /usr/lib/lib*.so.0.0.0
    FILE SIZE        VM SIZE
 --------------  --------------
   0.0%       0 -285.4% -1.82Gi    .bss
  45.6%  1.14Gi 179.3%  1.14Gi    .text
  15.6%   398Mi  61.1%   398Mi    .data
   9.5%   243Mi  37.4%   243Mi    .dynstr
   8.8%   225Mi  34.5%   225Mi    .rodata
   4.7%   121Mi  18.6%   121Mi    .rel.dyn
   4.3%   109Mi  16.7%   109Mi    .eh_frame
   2.8%  70.6Mi  10.8%  70.4Mi    .dynsym
   1.8%  47.1Mi   7.2%  46.9Mi    .data.rel.ro
   1.0%  25.4Mi   3.9%  25.2Mi    .gnu.hash
   1.0%  25.1Mi   3.8%  24.9Mi    .plt
   0.9%  23.6Mi   3.6%  23.4Mi    .gcc_except_table
   0.9%  22.5Mi   0.0%       0    .gnu_debugdata
   0.7%  18.5Mi   2.8%  18.3Mi    .eh_frame_hdr
   0.6%  14.4Mi   1.4%  8.80Mi    [39 Others]
   0.5%  12.6Mi   1.9%  12.4Mi    .rel.plt
   0.4%  9.69Mi   0.0%       0    [Unmapped]
   0.4%  8.98Mi   1.4%  8.80Mi    .gnu.version
   0.3%  6.43Mi   1.0%  6.25Mi    .got.plt
   0.2%  5.26Mi   0.0%       0    .debug_info
   0.2%  4.01Mi   0.0%       0    .debug_str
 100.0% -1.50Gi 100.0%   651Mi    TOTAL

After this change:

$ bloaty /usr/lib/lib*.so.0.0.0
    FILE SIZE        VM SIZE
 --------------  --------------
   0.0%       0  47.1%  2.18Gi    .bss
  45.6%  1.14Gi  24.6%  1.14Gi    .text
  15.6%   398Mi   8.4%   398Mi    .data
   9.5%   243Mi   5.1%   243Mi    .dynstr
   8.8%   225Mi   4.7%   225Mi    .rodata
   4.7%   121Mi   2.5%   121Mi    .rel.dyn
   4.3%   109Mi   2.3%   109Mi    .eh_frame
   2.8%  70.6Mi   1.5%  70.4Mi    .dynsym
   1.8%  47.1Mi   1.0%  46.9Mi    .data.rel.ro
   1.0%  25.4Mi   0.5%  25.2Mi    .gnu.hash
   1.0%  25.1Mi   0.5%  24.9Mi    .plt
   0.9%  23.6Mi   0.5%  23.4Mi    .gcc_except_table
   0.9%  22.5Mi   0.0%       0    .gnu_debugdata
   0.7%  18.5Mi   0.4%  18.3Mi    .eh_frame_hdr
   0.6%  14.4Mi   0.2%  8.80Mi    [39 Others]
   0.5%  12.6Mi   0.3%  12.4Mi    .rel.plt
   0.4%  9.69Mi   0.0%       0    [Unmapped]
   0.4%  8.98Mi   0.2%  8.80Mi    .gnu.version
   0.3%  6.43Mi   0.1%  6.25Mi    .got.plt
   0.2%  5.26Mi   0.0%       0    .debug_info
   0.2%  4.01Mi   0.0%       0    .debug_str
 100.0%  2.50Gi 100.0%  4.64Gi    TOTAL

+++ src/bloaty.cc
@@ -451,7 +451,7 @@ class Rollup {
     }
   }
 
-  static double Percent(ssize_t part, size_t whole) {
+  static double Percent(int64_t part, int64_t whole) {
     if (whole == 0) {
       if (part == 0) {
         return NAN;
@@ -674,7 +674,7 @@ std::string DoubleStringPrintf(const char *fmt, double d) {
   return std::string(buf);
 }
 
-std::string SiPrint(ssize_t size, bool force_sign) {
+std::string SiPrint(int64_t size, bool force_sign) {
   const char *prefixes[] = {"", "Ki", "Mi", "Gi", "Ti"};
   size_t num_prefixes = 5;
   size_t n = 0;
@@ -687,7 +687,7 @@ std::string SiPrint(ssize_t size, bool force_sign) {
   std::string ret;
 
   if (fabs(size_d) > 100 || n == 0) {
-    ret = std::to_string(static_cast<ssize_t>(size_d)) + prefixes[n];
+    ret = std::to_string(static_cast<int64_t>(size_d)) + prefixes[n];
     if (force_sign && size > 0) {
       ret = "+" + ret;
     }
